+++ Runtime/CyberiadaGraph/CyberiadaGraphDocument.cs
@@ -1,0 +1,10 @@
+﻿namespace Talent.Graphs
+{
+    public class CyberiadaGraphDocument
+    {
+        public CyberiadaGraph RootGraph { get; set; }
+        public string Target { get; set; }
+        public string Name { get; set; }
+        public string ReferenceGraphId { get; set; }
+    }
+}