+++ Runtime/CyberiadaGraph/GraphData.cs
@@ -1,17 +1,10 @@
 namespace Talent.Graphs
 {
     public class GraphData : IClonable<GraphData>
-    {
-        public string Name { get; set; }
-        public string ReferenceGraphID { get; set; }
-
-        /// <summary>
-        /// Creates a copy of the graph data
-        /// </summary>
+    { 
         public GraphData GetCopy()
         {
-            GraphData resultData = new GraphData { Name = Name };
-
+            GraphData resultData = new GraphData();
             return resultData;
         }
     }