+++ Runtime/CyberiadaGraph/Node.cs
@@ -1,5 +1,3 @@
-using System.Text;
-
 namespace Talent.Graphs
 {
     /// <summary>
@@ -65,25 +63,6 @@ namespace Talent.Graphs
             }
 
             return resultNode;
-        }
-
-        /// <summary>
-        /// Custom ToString realization for creating more representive string visualization of node data
-        /// </summary>
-        /// <returns></returns>
-        public override string ToString()
-        {
-            StringBuilder stringBuilder = new();
-
-            stringBuilder.AppendLine($"NODE({ID})({Data})");
-            stringBuilder.AppendLine($"{nameof(ParentNode)}={ParentNode?.ID}");
-
-            if (NestedGraph != null)
-            {
-                stringBuilder.AppendLine($"{NestedGraph}");
-            }
-
-            return stringBuilder.ToString();
         }
     }
 }