+++ Runtime/CyberiadaGraph/NodeData.cs
@@ -19,7 +19,7 @@ namespace Talent.Graphs
         /// <summary>
         /// Events of node
         /// </summary>
-        public IEnumerable<Event> Events => _events;
+        public IReadOnlyList<Event> Events => _events;
 
         public NodeData(string vertex = "")
         {
@@ -66,20 +66,6 @@ namespace Talent.Graphs
         }
 
         #endregion
-
-        public override string ToString()
-        {
-            StringBuilder stringBuilder = new();
-
-            stringBuilder.AppendLine($"NodeData({VisualData.Name})({VisualData.Position})");
-
-            foreach (Event @event in _events)
-            {
-                stringBuilder.AppendLine($"{@event}\n");
-            }
-
-            return stringBuilder.ToString();
-        }
     }
 
     public class NodeVisualData